fix: resolve typecheck errors across skills, RAG, and chat modules
