+++ packages/blog/server/api/admin/config.get.ts
@@ -1,5 +1,3 @@
-import { envSchema, getMaskedConfig } from '../../utils/env-config'
-
 export default defineEventHandler(async (event) => {
   const session = await getUserSession(event)
   if (!session.user) {
