+++ packages/blog/server/api/chatbots/index.get.ts
@@ -1,5 +1,3 @@
-import { eq, useDrizzle, tables } from '../../utils/drizzle'
-
 defineRouteMeta({
   openAPI: {
     description: 'List all active chatbots with persona and theme info',
