+++ packages/blog/server/api/knowledge-bases/index.get.ts
@@ -1,5 +1,3 @@
-import { capabilityRegistry } from '../../utils/capabilities'
-
 defineRouteMeta({
   openAPI: {
     description: 'List all available knowledge bases',
