+++ packages/blog/server/utils/capabilities/builtin/blog-assistant.ts
@@ -1,5 +1,3 @@
-import type { Capability } from '../types'
-
 export const blogAssistantCapability: Capability = {
   slug: 'blog-assistant',
   name: 'Blog Assistant',
