--- packages/blog/server/utils/capabilities/index.ts
@@ -1,8 +1,0 @@
-// Main exports for capabilities system
-export * from './types'
-export * from './registry'
-export * from './personas'
-export * from './knowledge-bases'
-
-// Re-export built-in capabilities
-export { builtInCapabilities } from './builtin'
